Support output files in current directory without './' (#572)
* Support output files in current directory without './'

convert_dir_to_note_sequences.py doesn't support specifying an output file in the current directory unless the user prefixes it with './' or uses the full path. This is because the script tries to create the output dir '' (empty string) which fails. This fixes the problem by not creating an output_dir if the output_file doesn't contain a directory component. 

Example command that is fixed by this change:
$ convert_dir_to_note_sequences --input_dir=lmd_matched --output_file output.tfrecord --recursive

* use tf.gfile.MakeDirs for internal compatibility