Propagate instrument name from and to midi file
Use the InstrumentInfo message in NoteSequence in order to propagate instrument name from and to prettyMidi. The instrument name is important for midi channel  coherency between NoteSequence and midi files. It is necessary for example for the case when midi is taken from DAW for processing and then send to the DAW again with the same midi channel setup